+++ bin/command.go
@@ -23,13 +23,15 @@ type SubCmdAlias struct {
 
 // 命令结构体
 type Command struct {
-	App App
-	SCA *SubCmdAlias
+	App  App
+	SCA  *SubCmdAlias
+	Util *CmdUitl
 }
 
 // 引用初始化接口
 func (c *Command) Init() {
 	c.App = *app
+	c.Util = &CmdUitl{c}
 }
 
 // 入口/内部分发(Entrance)
@@ -65,4 +67,26 @@ func (c *Command) SubCommandUnfind(subCmd string) {
 // 帮助说明
 func (c *Command) Help() {
 	fmt.Println("  项目帮助说明，外部通过： $ help [访问] 来查看对应的命令帮助")
+}
+
+// Command 协助方法
+// 通过 cmdInst 与 命令程序解析
+type CmdUitl struct {
+	cmdInst *Command
+}
+
+// 二级命令别名
+func (cu *CmdUitl) BaseSubCAlias(inst interface{}, alias map[string][]string) *SubCmdAlias {
+	if alias == nil {
+		alias = map[string][]string{}
+	}
+	// 帮助程序
+	alias["help"] = []string{"h"}
+
+	// 获取二级对象
+	csa := &SubCmdAlias{
+		Alias: alias,
+		Self:  inst,
+	}
+	return csa
 }